% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fecundity.r
\name{fecundity}
\alias{fecundity}
\title{A title of your function}
\usage{
fecundity(L, a, b)
}
\arguments{
\item{L}{fish length}

\item{a}{constant a}

\item{b}{constant b}
}
\value{
This function returns fish fecundity
}
\description{
A short description about the function.
}
\examples{
fecundity(36, 0.2, 3)
}

